requirements.txt
# =============================================================================
# Lumen test dependencies
# =============================================================================

# Test framework
pytest>=7.0
pytest-xdist

# Lumen core (editable install handled by Dockerfile / setup.py)
# torch>=2.0
# triton

# Quantization reference (torchao MX-format / FP8 primitives)
torchao>=0.8

# AITER — installed from third_party/aiter (editable) in Dockerfile
# Megatron-LM — installed from git clone in Dockerfile (not on PyPI)

# HuggingFace (model/dataset download helpers)
huggingface_hub>=0.20
datasets>=2.16
transformers>=4.38

# LoRA (via HuggingFace PEFT)
peft>=0.9

# Build tools (needed for mori submodule compilation)
setuptools-scm

# Utilities
numpy>=1.26.0
